--- a/go.mod
+++ b/go.mod
@@ -1,12 +1,13 @@
 module go.dedis.ch/kyber/v3
 
 require (
+	github.com/cloudflare/circl v1.3.7
 	github.com/kilic/bls12-381 v0.1.0
 	github.com/stretchr/testify v1.3.0
 	go.dedis.ch/fixbuf v1.0.3
 	go.dedis.ch/protobuf v1.0.11
-	golang.org/x/crypto v0.0.0-20190123085648-057139ce5d2b
-	golang.org/x/sys v0.0.0-20201101102859-da207088b7d1
+	golang.org/x/crypto v0.17.0
+	golang.org/x/sys v0.15.0
 )
 
 go 1.13