--- a/group/curve25519/basic.go
+++ b/group/curve25519/basic.go
@@ -1,3 +1,4 @@
+//go:build experimental
 // +build experimental
 
 package curve25519