--- a/pairing/bls12381/kilic/scalar.go
+++ b/pairing/bls12381/kilic/scalar.go
@@ -1,4 +1,4 @@
-package bls12381
+package kilic
 
 import (
 	"math/big"
@@ -9,7 +9,6 @@ import (
 
 var curveOrder, _ = new(big.Int).SetString("73eda753299d7d483339d80809a1d80553bda402fffe5bfeffffffff00000001", 16)
 
-// NewKyberScalar returns a new scalar value
-func NewKyberScalar() kyber.Scalar {
+func NewScalar() kyber.Scalar {
 	return mod.NewInt64(0, curveOrder)
 }