--- a/pairing/bn256/gfp_decl.go
+++ b/pairing/bn256/gfp_decl.go
@@ -1,3 +1,4 @@
+//go:build (amd64 && !generic) || (arm64 && !generic)
 // +build amd64,!generic arm64,!generic
 
 package bn256