--- a/pairing/bn256/gfp_generic.go
+++ b/pairing/bn256/gfp_generic.go
@@ -1,3 +1,4 @@
+//go:build (!amd64 && !arm64) || generic
 // +build !amd64,!arm64 generic
 
 package bn256