@AnomalRoil 2024-03-10
Good point, feel free to do another pr to add a gnark_bls12381 package in the pairing folder 😬