@pierluca 2024-03-11
I seem to recall (I may be wrong here) that we'd discussed it would make sense to put this kind of extra libraries in separate repos. Otherwise it makes it impossible to include Kyber in a project without also depending on all these other extra (competing) dependencies.
Do I remember this incorrectly @matteosz @AnomalRoil ?
We could of course keep everything as-is for now, out of convenience, and just split them out once we're satisfied with the state of the APIs for a v4.