@matteosz 2024-03-11
> I seem to recall (I may be wrong here) that we'd discussed it would make sense to put this kind of extra libraries in separate repos. Otherwise it makes it impossible to include Kyber in a project without also depending on all these other extra (competing) dependencies. Do I remember this incorrectly @matteosz @AnomalRoil ? We could of course keep everything as-is for now, out of convenience, and just split them out once we're satisfied with the state of the APIs for a v4.

There was already an external repo for kilic made by drand, while the circl was integrated into the kyber drand fork, so the actual original idea was to bring both back into kyber.

My thoughts on that are that I don't see any strong reason for which increasing the dependencies is worse than creating separate repos, one for each extra dependency. This makes the project more fragmented, while still requiring maintenance on each of such separate repos. Besides, it's not a service dependency that we have, where each dependency represents a failure point for the service; those are just stable libraries frozen at some specific versions.

So my preference would go towards having everything centralized but it's just my opinion.