@AnomalRoil 2024-03-18
I guess my only worry with this PR is that it should probably be rebased on #452 and #487 rather than be giant merge commits.

In general I'd advice against every using merge commits except for very small changes or changes on a branch that will be merged back into the original branch (since then the original branch already has these commits).